fix nits and remove resource/upvote requesters
